Added persistence unit test, and OS X support for the Kit (but not example)
The persistence unit test, located in PersistenceTests.swift,
demonstrates a case where an NSManagedObject subclass can not be
restored from an archive on disk, because it lacks an
init(entity:insertIntoManagedObjectContext:) initializer.

In addition, support for OS X is added, although perhaps not in an
ideal way, just to check that it’s working. It probably contains some
info.plist entries and bundle identifiers etc. that may not be ideal
for the product.